backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sentence-transformers
faiss-cpu
PyMuPDF
numpy
python-dotenv
requests
pytest
